📝Feat: ReviewResponse, ReviewEntity에 reviewId 추가 (수정)

+++ src/main/kotlin/com/sparta/tobiro/domain/reservation/dto/ReviewResponse.kt
@@ -1,7 +1,7 @@
 package com.sparta.tobiro.domain.reservation.dto
 
 data class ReviewResponse (
-        val reviewId: Int,
+        val reviewId: Long,
         val memberName: String,
         val content: String,
         val rating: Int
